Reverted changes and made an API
I changed back to some previous things I had commented and also created an API that would allow future applications to access the generator. Going back to previous changes meant that the words aren't repeating as much as before and the API checks if the word has been used previously in the sentence.